requirements.txt
fastapi
uvicorn
langchain
llama-index
sentence-transformers
faiss-cpu
transformers
torch
langdetect
whisper
pydantic